+++ backend/src/main/java/com/backend/budgetboss/manualinstitution/ManualInstitutionServiceImpl.java
@@ -4,48 +4,53 @@ import com.backend.budgetboss.manualinstitution.dto.ManualInstitutionResponseDTO
 import com.backend.budgetboss.manualinstitution.dto.UpdateManualInstitutionDTO;
 import com.backend.budgetboss.manualinstitution.helper.ManualInstitutionHelper;
 import com.backend.budgetboss.user.User;
-import java.util.List;
+import com.backend.budgetboss.user.helper.UserHelper;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class ManualInstitutionServiceImpl implements ManualInstitutionService {
+    private final ManualInstitutionRepository manualInstitutionRepository;
+    private final ManualInstitutionHelper manualInstitutionHelper;
+    private final UserHelper userHelper;
+    private final ModelMapper modelMapper;
 
-  private final ManualInstitutionRepository manualInstitutionRepository;
-  private final ManualInstitutionHelper manualInstitutionHelper;
-  private final ModelMapper modelMapper;
+    public ManualInstitutionServiceImpl(ManualInstitutionRepository manualInstitutionRepository,
+                                        ManualInstitutionHelper manualInstitutionHelper,
+                                        UserHelper userHelper,
+                                        ModelMapper modelMapper) {
+        this.manualInstitutionRepository = manualInstitutionRepository;
+        this.manualInstitutionHelper = manualInstitutionHelper;
+        this.userHelper = userHelper;
+        this.modelMapper = modelMapper;
+    }
 
-  public ManualInstitutionServiceImpl(ManualInstitutionRepository manualInstitutionRepository,
-      ManualInstitutionHelper manualInstitutionHelper,
-      ModelMapper modelMapper) {
-    this.manualInstitutionRepository = manualInstitutionRepository;
-    this.manualInstitutionHelper = manualInstitutionHelper;
-    this.modelMapper = modelMapper;
-  }
+    @Override
+    public List<ManualInstitutionResponseDTO> getManualInstitutions() {
+        return manualInstitutionRepository.findByUser(userHelper.getUser())
+                .stream()
+                .map(manualInstitution -> modelMapper.map(manualInstitution, ManualInstitutionResponseDTO.class))
+                .toList();
+    }
 
-  @Override
-  public List<ManualInstitutionResponseDTO> getManualInstitutions(User user) {
-    return manualInstitutionRepository.findByUser(user)
-        .stream()
-        .map(manualInstitution -> modelMapper.map(manualInstitution,
-            ManualInstitutionResponseDTO.class))
-        .toList();
-  }
+    @Override
+    public ManualInstitutionResponseDTO updateManualInstitution(Long id, UpdateManualInstitutionDTO updateManualInstitutionDTO) {
+        User user = userHelper.getUser();
+        ManualInstitution manualInstitution = manualInstitutionHelper.getManualInstitution(id);
 
-  @Override
-  public ManualInstitutionResponseDTO updateManualInstitution(
-      User user,
-      Long id,
-      UpdateManualInstitutionDTO updateManualInstitutionDTO) {
-    ManualInstitution manualInstitution = manualInstitutionHelper.getByUserAndId(user, id);
-    manualInstitution.setName(updateManualInstitutionDTO.getName());
-    return modelMapper.map(manualInstitutionRepository.save(manualInstitution),
-        ManualInstitutionResponseDTO.class);
-  }
+        manualInstitutionHelper.assertManualInstitutionOwnership(user, manualInstitution);
+        manualInstitution.setName(updateManualInstitutionDTO.getName());
+        return modelMapper.map(manualInstitutionRepository.save(manualInstitution), ManualInstitutionResponseDTO.class);
+    }
 
-  @Override
-  public void deleteManualInstitution(User user, Long id) {
-    ManualInstitution manualInstitution = manualInstitutionHelper.getByUserAndId(user, id);
-    manualInstitutionRepository.delete(manualInstitution);
-  }
+    @Override
+    public void deleteManualInstitution(Long id) {
+        User user = userHelper.getUser();
+        ManualInstitution manualInstitution = manualInstitutionHelper.getManualInstitution(id);
+
+        manualInstitutionHelper.assertManualInstitutionOwnership(user, manualInstitution);
+        manualInstitutionRepository.delete(manualInstitution);
+    }
 }
